fix(ui): better link previews
